Fix intermittent crash on Windows
We were using `_tcsset_s` to zero out a buffer, which isn't a valid use
case -- it takes a buffer size, but it requires that buffer to already
be zero terminated. That meant that if the buffer didn't randomly
contain a `\0`, it would trigger the security buffer checks and crash.